+++ src/main/java/joshie/harvestmoon/util/IShippable.java
@@ -5,5 +5,5 @@ import net.minecraft.item.ItemStack;
 public interface IShippable {
     public static final double SELL_QUALITY_MODIFIER = 0.01825D;
     
-    public int getSellValue(ItemStack stack);
+    public long getSellValue(ItemStack stack);
 }
